requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
python-decouple